Refactor the Sampler class to decouple it from the implementation details of graph generators, models and sampling methods (now it won't need modifications if new graph generators are added, some assumptions regarding those three elements must still be respected).
Make the Gibbs sampling method a model class method. The sampling method will be passed by name to the Sampler class and must be a class method of the model we are sampling from.

Rename aux methods to graph generators (a collection of graph generator methods suited for our needs).